+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+- **scoop-status:** Check bucket status, improve output ([#5011](https://github.com/ScoopInstaller/Scoop/issues/5011))
 - **scoop-info:** Show app installed/download size ([#4886](https://github.com/ScoopInstaller/Scoop/issues/4886))
 
 ### Bug Fixes